chore: new tests to post endpoint
